+++ examples/multiprocessing/lock.py
@@ -19,7 +19,5 @@ if __name__ == "__main__":
     with Pool() as p:
         p.map_async(f, [[lock, q]] * n)
 
-    for _ in range(n):
-        print(q.get())
-    
-
+        for _ in range(n):
+            print(q.get())