+++ lithops/multiprocessing/reduction.py
@@ -7,6 +7,7 @@
 # Licensed to PSF under a Contributor Agreement.
 #
 # Modifications Copyright (c) 2020 Cloudlab URV
+#
 
 from abc import ABCMeta, abstractmethod
 import copyreg
@@ -19,20 +20,22 @@ import sys
 
 from . import context
 
-__all__ = ['send_handle', 'recv_handle', 'ForkingPickler', 'register', 'dump']
-
+__all__ = ['send_handle', 'recv_handle', 'ForkingPickler', 'register', 'dump', 'DefaultPickler']
 
 HAVE_SEND_HANDLE = (sys.platform == 'win32' or
                     (hasattr(socket, 'CMSG_LEN') and
                      hasattr(socket, 'SCM_RIGHTS') and
                      hasattr(socket.socket, 'sendmsg')))
 
+
 #
 # Pickler subclass
 #
 
 class ForkingPickler(pickle.Pickler):
-    '''Pickler subclass used by multiprocessing.'''
+    """
+    Pickler subclass used by multiprocessing
+    """
     _extra_reducers = {}
     _copyreg_dispatch_table = copyreg.dispatch_table
 
@@ -43,7 +46,9 @@ class ForkingPickler(pickle.Pickler):
 
     @classmethod
     def register(cls, type, reduce):
-        '''Register a reduce function for a type.'''
+        """
+        Register a reduce function for a type
+        """
         cls._extra_reducers[type] = reduce
 
     @classmethod
@@ -54,10 +59,14 @@ class ForkingPickler(pickle.Pickler):
 
     loads = pickle.loads
 
+
 register = ForkingPickler.register
 
+
 def dump(obj, file, protocol=None):
-    '''Replacement for pickle.dump() using ForkingPickler.'''
+    """
+    Replacement for pickle.dump() using ForkingPickler
+    """
     ForkingPickler(file, protocol).dump(obj)
 
 
@@ -66,6 +75,7 @@ class DefaultPickler:
     loads = pickle.loads
     dump = pickle.dump
     dumps = pickle.dumps
+
 
 #
 # Platform specific definitions
@@ -76,16 +86,22 @@ if sys.platform == 'win32':
     __all__ += ['DupHandle', 'duplicate', 'steal_handle']
     import _winapi
 
+
     def duplicate(handle, target_process=None, inheritable=False):
-        '''Duplicate a handle.  (target_process is a handle not a pid!)'''
+        """
+        Duplicate a handle.  (target_process is a handle not a pid!)
+        """
         if target_process is None:
             target_process = _winapi.GetCurrentProcess()
         return _winapi.DuplicateHandle(
             _winapi.GetCurrentProcess(), handle, target_process,
             0, inheritable, _winapi.DUPLICATE_SAME_ACCESS)
 
+
     def steal_handle(source_pid, handle):
-        '''Steal a handle from process identified by source_pid.'''
+        """
+        Steal a handle from process identified by source_pid
+        """
         source_process_handle = _winapi.OpenProcess(
             _winapi.PROCESS_DUP_HANDLE, False, source_pid)
         try:
@@ -96,17 +112,21 @@ if sys.platform == 'win32':
         finally:
             _winapi.CloseHandle(source_process_handle)
 
+
     def send_handle(conn, handle, destination_pid):
         '''Send a handle over a local connection.'''
         dh = DupHandle(handle, _winapi.DUPLICATE_SAME_ACCESS, destination_pid)
         conn.send(dh)
 
+
     def recv_handle(conn):
         '''Receive a handle over a local connection.'''
         return conn.recv().detach()
 
+
     class DupHandle(object):
         '''Picklable wrapper for a handle.'''
+
         def __init__(self, handle, access, pid=None):
             if pid is None:
                 # We just duplicate the handle in the current process and
@@ -146,6 +166,7 @@ else:
     # On MacOSX we should acknowledge receipt of fds -- see Issue14669
     ACKNOWLEDGE = sys.platform == 'darwin'
 
+
     def sendfds(sock, fds):
         '''Send an array of fds over an AF_UNIX socket.'''
         fds = array.array('i', fds)
@@ -153,6 +174,7 @@ else:
         sock.sendmsg([msg], [(socket.SOL_SOCKET, socket.SCM_RIGHTS, fds)])
         if ACKNOWLEDGE and sock.recv(1) != b'A':
             raise RuntimeError('did not receive acknowledgement of fd')
+
 
     def recvfds(sock, size):
         '''Receive an array of fds over an AF_UNIX socket.'''
@@ -169,7 +191,7 @@ else:
                                    len(ancdata))
             cmsg_level, cmsg_type, cmsg_data = ancdata[0]
             if (cmsg_level == socket.SOL_SOCKET and
-                cmsg_type == socket.SCM_RIGHTS):
+                    cmsg_type == socket.SCM_RIGHTS):
                 if len(cmsg_data) % a.itemsize != 0:
                     raise ValueError
                 a.frombytes(cmsg_data)
@@ -179,15 +201,18 @@ else:
             pass
         raise RuntimeError('Invalid data received')
 
+
     def send_handle(conn, handle, destination_pid):
         '''Send a handle over a local connection.'''
         with socket.fromfd(conn.fileno(), socket.AF_UNIX, socket.SOCK_STREAM) as s:
             sendfds(s, [handle])
 
+
     def recv_handle(conn):
         '''Receive a handle over a local connection.'''
         with socket.fromfd(conn.fileno(), socket.AF_UNIX, socket.SOCK_STREAM) as s:
             return recvfds(s, 1)[0]
+
 
     def DupFd(fd):
         '''Return a wrapper for an fd.'''
@@ -200,6 +225,7 @@ else:
         else:
             raise ValueError('SCM_RIGHTS appears not to be available')
 
+
 #
 # Try making some callable types picklable
 #
@@ -209,22 +235,32 @@ def _reduce_method(m):
         return getattr, (m.__class__, m.__func__.__name__)
     else:
         return getattr, (m.__self__, m.__func__.__name__)
+
+
 class _C:
     def f(self):
         pass
+
+
 register(type(_C().f), _reduce_method)
 
 
 def _reduce_method_descriptor(m):
     return getattr, (m.__objclass__, m.__name__)
+
+
 register(type(list.append), _reduce_method_descriptor)
 register(type(int.__add__), _reduce_method_descriptor)
 
 
 def _reduce_partial(p):
     return _rebuild_partial, (p.func, p.args, p.keywords or {})
+
+
 def _rebuild_partial(func, args, keywords):
     return functools.partial(func, *args, **keywords)
+
+
 register(functools.partial, _reduce_partial)
 
 #
@@ -233,26 +269,35 @@ register(functools.partial, _reduce_partial)
 
 if sys.platform == 'win32':
     def _reduce_socket(s):
-        from .resource_sharer import DupSocket
-        return _rebuild_socket, (DupSocket(s),)
+        raise NotImplementedError()
+        # from .resource_sharer import DupSocket
+        # return _rebuild_socket, (DupSocket(s),)
+
+
     def _rebuild_socket(ds):
         return ds.detach()
+
+
     register(socket.socket, _reduce_socket)
 
 else:
     def _reduce_socket(s):
         df = DupFd(s.fileno())
         return _rebuild_socket, (df, s.family, s.type, s.proto)
+
+
     def _rebuild_socket(df, family, type, proto):
         fd = df.detach()
         return socket.socket(family, type, proto, fileno=fd)
+
+
     register(socket.socket, _reduce_socket)
 
 
 class AbstractReducer(metaclass=ABCMeta):
-    '''Abstract base class for use in implementing a Reduction class
+    """Abstract base class for use in implementing a Reduction class
     suitable for use in replacing the standard reduction mechanism
-    used in multiprocessing.'''
+    used in multiprocessing."""
     ForkingPickler = ForkingPickler
     register = register
     dump = dump