+++ lithops/scripts/cli.py
@@ -24,7 +24,7 @@ import shutil
 import lithops
 from lithops.tests import print_help, run_tests
 from lithops.utils import setup_logger, verify_runtime_name
-from lithops.config import default_config, extract_storage_config,\
+from lithops.config import get_mode, default_config, extract_storage_config,\
     extract_serverless_config, extract_standalone_config,\
     extract_localhost_config
 from lithops.constants import CACHE_DIR, LITHOPS_TEMP_DIR, RUNTIMES_PREFIX,\
@@ -46,15 +46,23 @@ def lithops_cli():
 
 
 @lithops_cli.command('clean')
-@click.option('--mode', default=None, type=click.Choice([SERVERLESS, LOCALHOST, STANDALONE], case_sensitive=True),
-              help='execution mode')
 @click.option('--config', '-c', default=None, help='use json config file')
+@click.option('--mode', '-m', default=None,
+              type=click.Choice([SERVERLESS, LOCALHOST, STANDALONE], case_sensitive=True),
+              help='execution mode')
+@click.option('--backend', '-b', default=None, help='compute backend')
 @click.option('--debug', '-d', is_flag=True, help='debug mode')
-def clean(mode, config, debug):
+def clean(config, mode, backend, debug):
     log_level = 'INFO' if not debug else 'DEBUG'
     setup_logger(log_level)
     logger.info('Cleaning all Lithops information')
-    config = default_config(config)
+
+    mode = mode or get_mode(config)
+    config_ow = {'lithops': {'mode': mode}}
+    if backend:
+        config_ow[mode] = {'backend': backend}
+    config = default_config(config, config_ow)
+
     storage_config = extract_storage_config(config)
     internal_storage = InternalStorage(storage_config)
 
@@ -84,8 +92,12 @@ def clean(mode, config, debug):
 
 @lithops_cli.command('test')
 @click.option('--config', '-c', default=None, help='use json config file')
+@click.option('--mode', '-m', default=None,
+              type=click.Choice([SERVERLESS, LOCALHOST, STANDALONE], case_sensitive=True),
+              help='execution mode')
+@click.option('--backend', '-b', default=None, help='compute backend')
 @click.option('--debug', '-d', is_flag=True, help='debug mode')
-def test_function(config, debug):
+def test_function(config, mode, backend, debug):
 
     if debug:
         setup_logger(logging.DEBUG)
@@ -93,7 +105,7 @@ def test_function(config, debug):
     def hello(name):
         return 'Hello {}!'.format(name)
 
-    fexec = lithops.FunctionExecutor(config=config)
+    fexec = lithops.FunctionExecutor(config=config, mode=mode, backend=backend)
     fexec.call_async(hello, 'World')
     result = fexec.get_result()
     print()
@@ -107,18 +119,26 @@ def test_function(config, debug):
 @lithops_cli.command('verify')
 @click.option('--test', '-t', default='all', help='run a specific test, type "-t help" for tests list')
 @click.option('--config', '-c', default=None, help='use json config file')
-@click.option('--mode', default=None, type=click.Choice([SERVERLESS, LOCALHOST, STANDALONE], case_sensitive=True),
+@click.option('--mode', '-m', default=None,
+              type=click.Choice([SERVERLESS, LOCALHOST, STANDALONE], case_sensitive=True),
               help='execution mode')
+@click.option('--backend', '-b', default=None, help='compute backend')
 @click.option('--debug', '-d', is_flag=True, help='debug mode')
-def verify(test, config, mode, debug):
+def verify(test, config, mode, backend, debug):
     if debug:
         setup_logger(logging.DEBUG)
 
     if test == 'help':
         print_help()
     else:
-        run_tests(test, mode, config)
+        run_tests(test, config, mode, backend)
 
+
+# /---------------------------------------------------------------------------/
+#
+# lithops logs
+#
+# /---------------------------------------------------------------------------/
 
 @click.group('logs')
 @click.pass_context
@@ -165,6 +185,12 @@ def get(job_key):
         print(content_file.read())
 
 
+# /---------------------------------------------------------------------------/
+#
+# lithops runtime
+#
+# /---------------------------------------------------------------------------/
+
 @click.group('runtime')
 @click.pass_context
 def runtime(ctx):
@@ -173,37 +199,30 @@ def runtime(ctx):
 
 @runtime.command('create')
 @click.argument('name')
-@click.option('--mode', default=None, type=click.Choice([SERVERLESS, LOCALHOST, STANDALONE], case_sensitive=True),
-              help='execution mode')
+@click.option('--backend', '-b', default=None, help='compute backend')
 @click.option('--memory', default=None, help='memory used by the runtime', type=int)
 @click.option('--timeout', default=None, help='runtime timeout', type=int)
 @click.option('--config', '-c', default=None, help='use json config file')
-def create(name, mode, memory, timeout, config):
+def create(name, backend, memory, timeout, config):
+    """ Create a serverless runtime """
     setup_logger(logging.DEBUG)
-    config = default_config(config)
+    logger.info('Creating new lithops runtime: {}'.format(name))
+
+    mode = SERVERLESS 
+    config_ow = {'lithops': {'mode': mode}}
+    if backend:
+        config_ow[mode] = {'backend': backend}
+    config = default_config(config, config_ow)
+
     storage_config = extract_storage_config(config)
     internal_storage = InternalStorage(storage_config)
 
-    mode = config['lithops']['mode'] if not mode else mode
-    if mode == SERVERLESS:
-        compute_config = extract_serverless_config(config)
-        compute_handler = ServerlessHandler(compute_config, storage_config)
-        mem = memory if memory else compute_config['runtime_memory']
-        to = timeout if timeout else compute_config['runtime_timeout']
-        runtime_key = compute_handler.get_runtime_key(name, mem)
-        runtime_meta = compute_handler.create_runtime(name, mem, timeout=to)
-    elif mode == STANDALONE:
-        compute_config = extract_standalone_config(config)
-        compute_handler = StandaloneHandler(compute_config)
-        runtime_key = compute_handler.get_runtime_key(name)
-        runtime_meta = compute_handler.create_runtime(name)
-    elif mode == LOCALHOST:
-        compute_config = extract_localhost_config(config)
-        compute_handler = LocalhostHandler(compute_config)
-        runtime_key = compute_handler.get_runtime_key(name)
-        runtime_meta = compute_handler.create_runtime(name)
-    else:
-        raise Exception('Unknown execution mode {}'.format(mode))
+    compute_config = extract_serverless_config(config)
+    compute_handler = ServerlessHandler(compute_config, storage_config)
+    mem = memory if memory else compute_config['runtime_memory']
+    to = timeout if timeout else compute_config['runtime_timeout']
+    runtime_key = compute_handler.get_runtime_key(name, mem)
+    runtime_meta = compute_handler.create_runtime(name, mem, timeout=to)
 
     try:
         internal_storage.put_runtime_meta(runtime_key, runtime_meta)
@@ -215,10 +234,18 @@ def create(name, mode, memory, timeout, config):
 @click.argument('name')
 @click.option('--file', '-f', default=None, help='file needed to build the runtime')
 @click.option('--config', '-c', default=None, help='use json config file')
-def build(name, file, config):
+@click.option('--backend', '-b', default=None, help='compute backend')
+def build(name, file, config, backend):
+    """ build a serverless runtime. """
     verify_runtime_name(name)
     setup_logger(logging.DEBUG)
-    config = default_config(config)
+
+    mode = SERVERLESS
+    config_ow = {'lithops': {'mode': mode}}
+    if backend:
+        config_ow[mode] = {'backend': backend}
+    config = default_config(config, config_ow)
+
     storage_config = extract_storage_config(config)
     compute_config = extract_serverless_config(config)
     compute_handler = ServerlessHandler(compute_config, storage_config)
@@ -228,16 +255,24 @@ def build(name, file, config):
 @runtime.command('update')
 @click.argument('name')
 @click.option('--config', '-c', default=None, help='use json config file')
-def update(name, config):
+@click.option('--backend', '-b', default=None, help='compute backend')
+def update(name, config, backend):
+    """ Update a serverless runtime """
     verify_runtime_name(name)
     setup_logger(logging.DEBUG)
-    config = default_config(config)
+
+    mode = SERVERLESS
+    config_ow = {'lithops': {'mode': mode}}
+    if backend:
+        config_ow[mode] = {'backend': backend}
+    config = default_config(config, config_ow)
+
     storage_config = extract_storage_config(config)
     internal_storage = InternalStorage(storage_config)
     compute_config = extract_serverless_config(config)
     compute_handler = ServerlessHandler(compute_config, storage_config)
 
-    timeout = config['lithops']['runtime_timeout']
+    timeout = compute_config['runtime_memory']
     logger.info('Updating runtime: {}'.format(name))
 
     runtimes = compute_handler.list_runtimes(name)
@@ -255,10 +290,18 @@ def update(name, config):
 @runtime.command('delete')
 @click.argument('name')
 @click.option('--config', '-c', default=None, help='use json config file')
-def delete(name, config):
+@click.option('--backend', '-b', default=None, help='compute backend')
+def delete(name, config, backend):
+    """ delete a serverless runtime """
     verify_runtime_name(name)
     setup_logger(logging.DEBUG)
-    config = default_config(config)
+
+    mode = SERVERLESS
+    config_ow = {'lithops': {'mode': mode}}
+    if backend:
+        config_ow[mode] = {'backend': backend}
+    config = default_config(config, config_ow)
+
     storage_config = extract_storage_config(config)
     internal_storage = InternalStorage(storage_config)
     compute_config = extract_serverless_config(config)