+++ lithops/serverless/backends/code_engine/config.py
@@ -16,16 +16,19 @@
 
 import os
 import sys
-from lithops.utils import version_str
+import shutil
+import subprocess as sp
 
-RUNTIME_DEFAULT = {'3.5': 'cactusone/lithops-code-engine-v3.5',
-                   '3.6': 'cactusone/lithops-code-engine-v3.6',
-                   '3.7': 'cactusone/lithops-code-engine-v3.7',
-                   '3.8': 'cactusone/lithops-code-engine-v3.8'}
+from lithops.utils import version_str
+from lithops.version import __version__
+
+RUNTIME_NAME = 'lithops-codeengine'
+
+DOCKER_PATH = shutil.which('docker')
 
 RUNTIME_TIMEOUT_DEFAULT = 600  # Default: 600 seconds => 10 minutes
 RUNTIME_MEMORY_DEFAULT = 128  # Default memory: 256 MB
-MAX_CONCURRENT_WORKERS = 1200
+MAX_CONCURRENT_WORKERS = 250
 CPU_DEFAULT = 1  # default number of CPU
 
 DEFAULT_API_VERSION = 'codeengine.cloud.ibm.com/v1beta1'
@@ -35,30 +38,73 @@ DEFAULT_VERSION = "v1beta1"
 FH_ZIP_LOCATION = os.path.join(os.getcwd(), 'lithops_codeengine.zip')
 
 
+DEFAULT_DOCKERFILE = """
+RUN apt-get update && apt-get install -y \
+        zip \
+        && rm -rf /var/lib/apt/lists/*
+
+RUN pip install --upgrade setuptools six pip \
+    && pip install --no-cache-dir \
+        gunicorn==19.9.0 \
+        pika==0.13.1 \
+        flask \
+        gevent \
+        glob2 \
+        ibm-cos-sdk \
+        redis \
+        requests \
+        PyYAML \
+        kubernetes \
+        numpy
+
+ENV CONCURRENCY 4
+ENV TIMEOUT 600
+ENV PYTHONUNBUFFERED TRUE
+
+# Copy Lithops proxy and lib to the container image.
+ENV APP_HOME /lithops
+WORKDIR $APP_HOME
+
+COPY lithops_codeengine.zip .
+RUN unzip lithops_codeengine.zip && rm lithops_codeengine.zip
+"""
+
+
 def load_config(config_data):
+    if 'code_engine' not in config_data:
+        config_data['code_engine'] = {}
+
+    if 'runtime_cpu' not in config_data['code_engine']:
+        config_data['code_engine']['runtime_cpu'] = CPU_DEFAULT
+
     if 'runtime_memory' not in config_data['serverless']:
         config_data['serverless']['runtime_memory'] = RUNTIME_MEMORY_DEFAULT
     if 'runtime_timeout' not in config_data['serverless']:
         config_data['serverless']['runtime_timeout'] = RUNTIME_TIMEOUT_DEFAULT
+
     if 'runtime' not in config_data['serverless']:
-        python_version = version_str(sys.version_info)
-        try:
-            config_data['serverless']['runtime'] = RUNTIME_DEFAULT[python_version]
-        except KeyError:
-            raise Exception('Unsupported Python version: {}'.format(python_version))
+        if 'docker_user' not in config_data['code_engine']:
+            cmd = "{} info".format(DOCKER_PATH)
+            docker_user_info = sp.check_output(cmd, shell=True, encoding='UTF-8',
+                                               stderr=sp.STDOUT)
+            for line in docker_user_info.splitlines():
+                if 'Username' in line:
+                    _, useranme = line.strip().split(':')
+                    config_data['code_engine']['docker_user'] = useranme.strip()
+                    break
+
+        if 'docker_user' not in config_data['code_engine']:
+            raise Exception('You must provide "docker_user" param in config '
+                            'or execute "docker login"')
+
+        docker_user = config_data['code_engine']['docker_user']
+        python_version = version_str(sys.version_info).replace('.', '')
+        revision = 'latest' if 'dev' in __version__ else __version__.replace('.', '')
+        runtime_name = '{}/{}-v{}:{}'.format(docker_user, RUNTIME_NAME, python_version, revision)
+        config_data['serverless']['runtime'] = runtime_name
 
     config_data['serverless']['remote_invoker'] = True
 
     if 'workers' not in config_data['lithops'] or \
        config_data['lithops']['workers'] > MAX_CONCURRENT_WORKERS:
         config_data['lithops']['workers'] = MAX_CONCURRENT_WORKERS
-
-    if 'code_engine' in config_data:
-        if 'runtime_cpu' not in config_data['code_engine']:
-            config_data['code_engine']['runtime_cpu'] = CPU_DEFAULT
-        if 'api_version' not in config_data['code_engine']:
-            config_data['code_engine']['api_version'] = DEFAULT_API_VERSION
-        if 'group' not in config_data['code_engine']:
-            config_data['code_engine']['group'] = DEFAULT_GROUP
-        if 'version' not in config_data['code_engine']:
-            config_data['code_engine']['version'] = DEFAULT_VERSION